pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "boran-ai"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "requests",
    "chromadb",
    "sentence-transformers",
    "huggingface-hub",
    "pypdf",
    "pdf2image",
    "pillow",
    "pytesseract",
    "python-docx",
    "python-multipart",
    "networkx",
    "bcrypt",
    "PyJWT",
    "sqlalchemy",
    "psycopg[binary]"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx"
]

voice = [
    "faster-whisper",
    "edge-tts",
    "TTS"
]

[tool.setuptools.packages.find]
include = ["app*"]
